requirement.txt
langchain 
pinecone
langchain-community
pypdf
langchain_pinecone
streamlit